% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAUM.R
\name{BAUM}
\alias{BAUM}
\title{BAUM}
\usage{
BAUM(
  r,
  Q,
  g,
  num_samples = 5000,
  m0 = 0,
  G = 50,
  mu_1 = 1:50,
  a_1 = 10000,
  a_2 = 1000,
  a_3 = 2000,
  a_4 = 1000,
  a_5 = 50,
  b_1 = 1,
  b_3 = 1000,
  b_4 = 1,
  b_5 = 50,
  a_g = rep(10, 50),
  b_g = rep(10, 50),
  pi = c(0.5, 0.5),
  rho = c(0.1, 0.1)
)
}
\arguments{
\item{r}{a vector of length p containing feature-specific test statistics}

\item{Q}{a p times k matrix of matching uncertainty prior information}

\item{g}{an igraph object of network}

\item{num_samples}{a positive integer, number of MCMC samples}

\item{m0}{an integer, mean for the null component}

\item{G}{a positive integer, number of clusters for the approximation of Dirichlet process mixture (DPM) of the alternative component}

\item{mu_1}{a vector of length G, the prior means of Gaussian distribution in G clusters for the approximation of DPM, i.e., m_g for g = 1, ..., G}

\item{a_1}{a number, shape of the Inverse Gamma prior for gamma_0}

\item{a_2}{a number, shape of the Inverse Gamma prior for gamma_g}

\item{a_3}{a number, shape of the Inverse Gamma prior for sigma^2}

\item{a_4}{a number, shape of the Gamma prior for beta_g, i.e., the scale (rate of Gamma) for gamma_g}

\item{a_5}{a number, shape of the Inverse Gamma prior for sigma^2_g, i.e., the variance of m_g}

\item{b_1}{a number, rate of the Inverse Gamma prior for gamma_0}

\item{b_3}{a number, rate of the Inverse Gamma prior for sigma^2}

\item{b_4}{a number, rate of the Gamma prior for beta_g, i.e., the scale (rate of Gamma) for gamma_g}

\item{b_5}{a number, rate of the Inverse Gamma prior for sigma^2_g, i.e., the variance of m_g}

\item{a_g}{a vector of length G, Beta prior parameter for p_g for g = 1, .., G}

\item{b_g}{a vector of length G, Beta prior parameter for p_g for g = 1, .., G}

\item{pi}{a vector of length 2, prior label distribution; the two elements should be between 0 and 1, and sum up to 1}

\item{rho}{a vector of 2 positive numbers, global neighborhood parameters}
}
\value{
para_sample, a list of MCMC samples for each parameter, including m_1, beta, gamma_0, gamma_1, sigma_sq, sigma_sq_K, eta_0, eta_1, Lambda, z, N_g and K.
}
\description{
function for Bayesian Analysis for Untargeted Metabolomics data (BAUM)
}
\examples{
result <- custom_add(1,2)
}
