% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAUM.R
\name{FDR_threshold}
\alias{FDR_threshold}
\title{FDR_threshold}
\usage{
FDR_threshold(posterior_samples, alpha)
}
\arguments{
\item{posterior_samples}{a num_sampels times k matrix, the posterior samples of latent class labels}

\item{alpha}{a number between 0 and 1, the FDR cutoff}
}
\value{
the threshold on inclusion probabilities based on the input of FDR alpha
}
\description{
function for generating a threshold on inclusion probabilities based on the input of FDR alpha
}
