% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAUM.R
\name{FDR_threshold_from_prob}
\alias{FDR_threshold_from_prob}
\title{FDR_threshold_from_prob}
\usage{
FDR_threshold_from_prob(posterior_prob, alpha)
}
\arguments{
\item{posterior_prob}{a vector of length k, the posterior inclusion probabilities}

\item{alpha}{a number between 0 and 1, the FDR cutoff}
}
\value{
the threshold on inclusion probabilities based on the input of FDR alpha
}
\description{
function for generating a threshold on inclusion probabilities based on the input of FDR alpha
}
