% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAUM.R
\name{generate_hyperpara_grid}
\alias{generate_hyperpara_grid}
\title{generate_hyperpara_grid}
\usage{
generate_hyperpara_grid(
  ratio = c(0.5, 2),
  a_1 = 10000,
  a_2 = 1000,
  a_3 = 2000,
  a_4 = 1000,
  a_5 = 50,
  b_1 = 1,
  b_3 = 1000,
  b_4 = 1,
  b_5 = 50,
  beta_pg = 10
)
}
\arguments{
\item{ratio}{a vector of 2 positive numbers (r1, r2). Each input parameter is multiplied by r1 and r2 as the value on the parameter grid.}

\item{a_1}{a number, shape of the Inverse Gamma prior for gamma_0}

\item{a_2}{a number, shape of the Inverse Gamma prior for gamma_g}

\item{a_3}{a number, shape of the Inverse Gamma prior for sigma^2}

\item{a_4}{a number, shape of the Gamma prior for beta_g, i.e., the scale (rate of Gamma) for gamma_g}

\item{a_5}{a number, shape of the Inverse Gamma prior for sigma^2_g, i.e., the variance of m_g}

\item{b_1}{a number, rate of the Inverse Gamma prior for gamma_0}

\item{b_3}{a number, rate of the Inverse Gamma prior for sigma^2}

\item{b_4}{a number, rate of the Gamma prior for beta_g, i.e., the scale (rate of Gamma) for gamma_g}

\item{b_5}{a number, rate of the Inverse Gamma prior for sigma^2_g, i.e., the variance of m_g}

\item{beta_pg}{a number, Beta prior parameter for p_g for g = 1, .., G}
}
\value{
grid, a matrix of parameter grid, each row is a parameter combination
}
\description{
function to generate a hyperpara grid for sensitivity analysis
}
