% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAUM.R
\name{p2fdr}
\alias{p2fdr}
\title{p2fdr}
\usage{
p2fdr(posterior_samples)
}
\arguments{
\item{posterior_samples}{a num_samples times k matrix, the posterior samples of latent class labels}
}
\value{
a vector of k, the FDR associated with each of k elements
}
\description{
function converting posterior inclusion probabilities to FDR
}
